feat(agent-runtime): scaffold @openape/agent-runtime + move execution cluster
