+++ packages/agent-runtime/src/agent-tools/bash.ts
