+++ packages/agent-runtime/src/index.ts
@@ -1,0 +1,20 @@
+export { runLoop, RpcSessionMap } from './agent-runtime.js'
+export type { ChatMessage, RunOptions, RunResult, RuntimeConfig, RunStreamHandlers, TraceEntry } from './agent-runtime.js'
+export { taskTools, TOOLS } from './agent-tools/index.js'
+export type { ToolDefinition } from './agent-tools/index.js'
+export { runApeShell } from './agent-tools/ape-shell-exec.js'
+export type { ApeShellResult } from './agent-tools/ape-shell-exec.js'
+// Coding cluster — used by apes commands/agents/code and the plan/review pipeline
+export { runCodingTask } from './coding/coding-loop.js'
+export type { CodingTaskInput, CodingTaskDeps, CodingTaskResult, ShellResult, ShellFn, MergeOutcome } from './coding/coding-loop.js'
+export { detectForge, buildIssueGet, buildPrCreate, buildPrMerge, buildPrStatus, registerForge, listForges, shq } from './coding/forge.js'
+export type { Forge, ForgeAdapter, PrCreateInput, PrMergeInput } from './coding/forge.js'
+export type { IssueRef, BranchNaming, TaskFraming } from './coding/issue-task.js'
+export { buildBranchName, buildTaskPrompt, slugify } from './coding/issue-task.js'
+export { createLlmReviewer, createLlmRiskAssessor } from './coding/llm-review.js'
+export { resolveMergePolicy, deriveRiskGlobs, extractWorkflowPaths, parseCodeowners } from './coding/derive-policy.js'
+export { decideMerge, classifyChange, SECURE_DEFAULT_POLICY, globToRegExp, matchesAny, loadMergePolicy } from './coding/merge-policy.js'
+export type { MergeDecision, MergePolicy, RiskAssessorFn, AgentRiskAssessment } from './coding/merge-policy.js'
+export { gateMerge } from './coding/review-gate.js'
+export type { ReviewerFn, ReviewVerdict } from './coding/review-gate.js'
+export { BudgetTracker, BudgetExceededError } from './coding/budget.js'
