+++ packages/apes/src/commands/agents/run.ts
@@ -4,9 +4,8 @@ import { join } from 'node:path'
 import { defineCommand } from 'citty'
 import consola from 'consola'
 import { CliError } from '../../errors'
-import { taskTools } from '../../lib/agent-tools'
-import { runLoop  } from '../../lib/agent-runtime'
-import type { RuntimeConfig } from '../../lib/agent-runtime'
+import { taskTools, runLoop } from '@openape/agent-runtime'
+import type { RuntimeConfig } from '@openape/agent-runtime'
 import { materializeSecrets } from '../../lib/agent-secrets-runtime'
 import { resolveTroopUrl, TroopClient  } from '../../lib/troop-client'
 import type { TaskSpec } from '../../lib/troop-client'
